+++ src/services/wallets/fund-distribution/service.spec.ts
@@ -480,141 +480,6 @@ describe('processDistributionCycle', () => {
 	});
 });
 
-describe('reconcileInFlightRequests', () => {
-	const inFlight = (transaction: Record<string, unknown>, overrides: Record<string, unknown> = {}) => ({
-		id: 'req-1',
-		batchId: 'batch-1',
-		amount: 20_000_000n,
-		assetUnit: 'lovelace',
-		targetWalletId: 'w1',
-		transactionId: 'tx-1',
-		TargetWallet: { walletAddress: 'addr_target_w1', deletedAt: null },
-		FundWallet: {
-			id: 'fund-1',
-			walletAddress: 'addr_fund',
-			paymentSourceId: 'ps-1',
-			deletedAt: null,
-			FundDistributionConfig: { enabled: true },
-			PaymentSource: { network: 'Preprod', deletedAt: null },
-		},
-		Transaction: {
-			txHash: null,
-			intendedTxHash: 'intended-1',
-			status: TransactionStatus.Pending,
-			createdAt: new Date(),
-			...transaction,
-		},
-		...overrides,
-	});
-
-	it('adopts a reconciliation-promoted batch as Submitted', async () => {
-		// reconcileOne promoted intendedTxHash -> txHash, meaning the tx IS on
-		// chain. Re-sending would double-spend; the batch must move to Submitted
-		// so the confirm phase picks it up.
-		phaseRows['in-flight'] = [inFlight({ txHash: 'abc' })];
-
-		await fundDistributionService.processDistributionCycle();
-
-		expect(mockRequestUpdateMany).toHaveBeenCalledWith({
-			where: { id: { in: ['req-1'] }, status: FundDistributionStatus.Pending },
-			data: { status: FundDistributionStatus.Submitted, txHash: 'abc' },
-		});
-		expect(mockQueueSent).toHaveBeenCalledWith(
-			expect.any(Object),
-			expect.objectContaining({
-				txHash: 'abc',
-				batchId: 'batch-1',
-				distributions: [expect.objectContaining({ assetUnit: 'lovelace' })],
-			}),
-			'ps-1',
-		);
-	});
-
-	it('releases a rolled-back batch so it can be rebuilt', async () => {
-		// TTL provably elapsed: the body can never land, so a fresh build with new
-		// inputs is safe.
-		phaseRows['in-flight'] = [inFlight({ status: TransactionStatus.RolledBack })];
-
-		await fundDistributionService.processDistributionCycle();
-
-		expect(mockRequestUpdateMany).toHaveBeenCalledWith({
-			where: { id: { in: ['req-1'] }, status: FundDistributionStatus.Pending },
-			data: { fundWalletId: null, transactionId: null, batchId: null },
-		});
-	});
-
-	it('fails a rolled-back batch instead of rebuilding toward an inactive target', async () => {
-		phaseRows['in-flight'] = [
-			inFlight(
-				{ status: TransactionStatus.RolledBack },
-				{ TargetWallet: { walletAddress: 'addr_target_w1', deletedAt: new Date() } },
-			),
-		];
-
-		await fundDistributionService.processDistributionCycle();
-
-		expect(mockRequestUpdateMany).toHaveBeenCalledWith({
-			where: { id: { in: ['req-1'] }, status: FundDistributionStatus.Pending },
-			data: {
-				status: FundDistributionStatus.Failed,
-				error: 'Distribution cancelled because its payment source or wallet is inactive',
-				transactionId: null,
-			},
-		});
-	});
-
-	it('releases a RolledBack batch even when it still carries a txHash', async () => {
-		// tx-sync marks a Transaction RolledBack BY txHash and leaves the hash set.
-		// Checking txHash first would promote a batch the chain already discarded,
-		// stranding it until the 30min confirm timeout for no reason.
-		phaseRows['in-flight'] = [inFlight({ txHash: 'abc', status: TransactionStatus.RolledBack })];
-
-		await fundDistributionService.processDistributionCycle();
-
-		expect(mockRequestUpdateMany).toHaveBeenCalledWith(
-			expect.objectContaining({ data: { fundWalletId: null, transactionId: null, batchId: null } }),
-		);
-		expect(mockRequestUpdateMany).not.toHaveBeenCalledWith(
-			expect.objectContaining({ data: expect.objectContaining({ status: FundDistributionStatus.Submitted }) }),
-		);
-	});
-
-	it('releases a never-broadcast orphan so its target is not stranded forever', async () => {
-		// Crash between creating the Transaction and recording intendedTxHash:
-		// nothing was signed, so nothing can be on chain. No other worker recovers
-		// this — reconciliation only considers rows WITH an intendedTxHash. Left
-		// linked, the row is excluded from rebuild AND counted as "already queued"
-		// by the scan, so the target would never be topped up again.
-		phaseRows['in-flight'] = [
-			inFlight({ txHash: null, intendedTxHash: null, createdAt: new Date(Date.now() - 3_600_000) }),
-		];
-
-		await fundDistributionService.processDistributionCycle();
-
-		expect(mockRequestUpdateMany).toHaveBeenCalledWith({
-			where: { id: { in: ['req-1'] }, status: FundDistributionStatus.Pending },
-			data: { fundWalletId: null, transactionId: null, batchId: null },
-		});
-	});
-
-	it('leaves a fresh never-broadcast row alone until the lock timeout passes', async () => {
-		// It may simply still be mid build/sign.
-		phaseRows['in-flight'] = [inFlight({ txHash: null, intendedTxHash: null, createdAt: new Date() })];
-
-		await fundDistributionService.processDistributionCycle();
-
-		expect(mockRequestUpdateMany).not.toHaveBeenCalled();
-	});
-
-	it('leaves a signed, in-flight batch alone', async () => {
-		phaseRows['in-flight'] = [inFlight({ txHash: null, intendedTxHash: 'intended-1' })];
-
-		await fundDistributionService.processDistributionCycle();
-
-		expect(mockRequestUpdateMany).not.toHaveBeenCalled();
-	});
-});
-
 // A second fund wallet on the same source, distinct id/address.
 const fundWallet2 = { ...fundWallet, id: 'fund-2', walletAddress: 'addr_fund_2' };
 
@@ -734,151 +599,5 @@ describe('batched dispatch', () => {
 
 		expect(mockProcessRequests).toHaveBeenCalledTimes(1);
 		expect(mockProcessRequests).toHaveBeenCalledWith(fundWallet, expect.any(Array));
-	});
-});
-
-describe('confirmSubmittedRequests', () => {
-	const submitted = (overrides: Record<string, unknown> = {}) => ({
-		id: 'req-1',
-		txHash: 'tx-hash-1',
-		updatedAt: new Date(Date.now() - 600_000),
-		fundWalletId: 'fund-1',
-		batchId: 'batch-1',
-		amount: 20_000_000n,
-		assetUnit: 'lovelace',
-		targetWalletId: 'w1',
-		transactionId: 'tx-1',
-		TargetWallet: { walletAddress: 'addr_target_w1' },
-		FundWallet: {
-			id: 'fund-1',
-			walletAddress: 'addr_fund',
-			lockedAt: new Date(),
-			pendingTransactionId: 'tx-1',
-			PaymentSource: { id: 'ps-1', PaymentSourceConfig: { rpcProviderApiKey: 'key' }, network: 'Preprod' },
-		},
-		...overrides,
-	});
-
-	it('only looks at rows old enough for the indexer to have seen', async () => {
-		phaseRows.submitted = [submitted()];
-
-		await fundDistributionService.processDistributionCycle();
-
-		// Without the age filter a tx submitted this cycle looks not-found and is
-		// marked Failed.
-		expect(mockRequestFindMany).toHaveBeenCalledWith(
-			expect.objectContaining({
-				where: expect.objectContaining({
-					status: FundDistributionStatus.Submitted,
-					txHash: { not: null },
-					updatedAt: { lt: expect.any(Date) },
-				}),
-			}),
-		);
-	});
-
-	it('confirms an in-flight batch after the last fund wallet is disabled', async () => {
-		phaseRows.submitted = [submitted()];
-		mockHotWalletFindMany.mockResolvedValue([]);
-		mockLookupChainTx.mockResolvedValue('found');
-
-		await fundDistributionService.processDistributionCycle();
-
-		expect(mockQueueConfirmed).toHaveBeenCalledWith(
-			expect.any(Object),
-			expect.objectContaining({
-				txHash: 'tx-hash-1',
-				distributions: [expect.objectContaining({ assetUnit: 'lovelace' })],
-			}),
-			'ps-1',
-		);
-	});
-
-	it('confirms a batch found on chain, touching only its own rows', async () => {
-		phaseRows.submitted = [submitted()];
-		mockLookupChainTx.mockResolvedValue('found');
-
-		await fundDistributionService.processDistributionCycle();
-
-		// Assert the WHERE, not just the data: an unscoped update would mark every
-		// distribution in the table Confirmed and a data-only assertion would pass.
-		expect(mockRequestUpdateMany).toHaveBeenCalledWith({
-			where: { id: { in: ['req-1'] }, status: FundDistributionStatus.Submitted },
-			data: { status: FundDistributionStatus.Confirmed, error: null },
-		});
-		expect(mockQueueConfirmed).toHaveBeenCalledWith(
-			expect.any(Object),
-			expect.objectContaining({ txHash: 'tx-hash-1' }),
-			'ps-1',
-		);
-	});
-
-	it('fails a batch absent from chain past the timeout, touching only its own rows', async () => {
-		phaseRows.submitted = [submitted({ updatedAt: new Date(Date.now() - 3_600_000) })];
-		mockLookupChainTx.mockResolvedValue('not-found');
-
-		await fundDistributionService.processDistributionCycle();
-
-		expect(mockRequestUpdateMany).toHaveBeenCalledWith({
-			where: { id: { in: ['req-1'] }, status: FundDistributionStatus.Submitted },
-			data: { status: FundDistributionStatus.Failed, error: 'Transaction not found on-chain after timeout' },
-		});
-		expect(mockQueueFailed).toHaveBeenCalledWith(
-			expect.any(Object),
-			expect.objectContaining({ txHash: 'tx-hash-1' }),
-			'ps-1',
-		);
-	});
-
-	it('keeps a not-yet-indexed batch Submitted and the wallet locked', async () => {
-		phaseRows.submitted = [submitted({ updatedAt: new Date(Date.now() - 600_000) })];
-		mockLookupChainTx.mockResolvedValue('not-found');
-
-		await fundDistributionService.processDistributionCycle();
-
-		// Within the confirmation timeout a 404 just means Blockfrost lags.
-		expect(mockQueueFailed).not.toHaveBeenCalled();
-		expect(mockHotWalletUpdateMany).not.toHaveBeenCalled();
-	});
-
-	it('does not fail or unlock a batch when the indexer is unhealthy', async () => {
-		phaseRows.submitted = [submitted({ updatedAt: new Date(Date.now() - 3_600_000) })];
-		mockLookupChainTx.mockResolvedValue('transient-error');
-
-		await fundDistributionService.processDistributionCycle();
-
-		// Inferring "not on chain" from a 5xx is how a landed tx gets marked Failed
-		// and re-sent. Even past the timeout, a transient error must not fail it.
-		expect(mockRequestUpdateMany).not.toHaveBeenCalled();
-		expect(mockHotWalletUpdateMany).not.toHaveBeenCalled();
-	});
-
-	it('releases only the lock held by the batch it resolved', async () => {
-		phaseRows.submitted = [submitted()];
-		mockLookupChainTx.mockResolvedValue('found');
-
-		await fundDistributionService.processDistributionCycle();
-
-		// Clearing a lock we do not own lets the next cycle rebuild rows whose tx
-		// is still in flight.
-		expect(mockHotWalletUpdateMany).toHaveBeenCalledWith({
-			where: { id: 'fund-1', pendingTransactionId: { in: ['tx-1'] } },
-			data: { lockedAt: null, pendingTransactionId: null },
-		});
-		expect(mockTransactionUpdateMany).toHaveBeenCalledWith({
-			where: { id: { in: ['tx-1'] }, status: TransactionStatus.Pending },
-			data: { status: TransactionStatus.Confirmed, lastCheckedAt: expect.any(Date) },
-		});
-	});
-
-	it('does not emit or unlock when another replica won the terminal transition', async () => {
-		phaseRows.submitted = [submitted()];
-		mockLookupChainTx.mockResolvedValue('found');
-		mockRequestUpdateMany.mockResolvedValue({ count: 0 });
-
-		await fundDistributionService.processDistributionCycle();
-
-		expect(mockQueueConfirmed).not.toHaveBeenCalled();
-		expect(mockHotWalletUpdateMany).not.toHaveBeenCalled();
 	});
 });
