+++ src/services/wallets/low-balance/service.spec.ts
@@ -211,19 +211,6 @@ describe('WalletLowBalanceMonitorService', () => {
 	});
 
 	const balanceMap = (amount: bigint) => new Map<string, bigint>([['lovelace', amount]]);
-	const createRuleRecord = (status: (typeof LowBalanceStatus)[keyof typeof LowBalanceStatus], enabled = true) => ({
-		id: 'rule-1',
-		hotWalletId: 'wallet-1',
-		assetUnit: 'lovelace',
-		thresholdAmount: 5000000n,
-		enabled,
-		topupEnabled: false,
-		topupAmount: null,
-		status,
-		lastKnownAmount: null,
-		lastCheckedAt: null,
-		lastAlertedAt: null,
-	});
 	const createBalanceFetchWallet = () => ({
 		id: 'wallet-1',
 		walletAddress: 'addr_test1...',
@@ -410,214 +397,6 @@ describe('WalletLowBalanceMonitorService', () => {
 		expect(mockLogWarn).not.toHaveBeenCalled();
 		expect(mockTriggerWalletLowBalance).not.toHaveBeenCalled();
 		expect(mockRecordWalletLowBalanceAlert).not.toHaveBeenCalled();
-	});
-
-	it('creates a rule and silently seeds low state when current balance is already below threshold', async () => {
-		mockCreateLowBalanceRule.mockResolvedValue(createRuleRecord(LowBalanceStatus.Unknown));
-		mockHotWalletFindFirst
-			.mockResolvedValueOnce(createBalanceFetchWallet())
-			.mockResolvedValueOnce(createWallet(LowBalanceStatus.Unknown));
-		mockFetchAddressBalanceMap.mockResolvedValue(balanceMap(4000000n));
-		mockFindUniqueLowBalanceRule.mockResolvedValue({
-			...createRuleRecord(LowBalanceStatus.Low),
-			lastKnownAmount: 4000000n,
-			lastCheckedAt: new Date('2026-03-10T12:00:00.000Z'),
-			lastAlertedAt: null,
-		});
-
-		const createdRule = await service.createRuleForWallet({
-			hotWalletId: 'wallet-1',
-			assetUnit: 'lovelace',
-			thresholdAmount: 5000000n,
-			enabled: true,
-		});
-
-		expect(mockCreateLowBalanceRule).toHaveBeenCalledWith(
-			expect.objectContaining({
-				data: expect.objectContaining({
-					status: LowBalanceStatus.Unknown,
-					lastKnownAmount: null,
-					lastCheckedAt: null,
-					lastAlertedAt: null,
-				}),
-			}),
-		);
-		expect(mockFetchAddressBalanceMap).toHaveBeenCalledWith({
-			network: Network.Preprod,
-			rpcProviderApiKey: 'provider-key',
-			address: 'addr_test1...',
-		});
-		expect(mockUpdateMany).toHaveBeenCalledWith(
-			expect.objectContaining({
-				where: expect.objectContaining({
-					id: 'rule-1',
-					status: LowBalanceStatus.Unknown,
-				}),
-				data: expect.objectContaining({
-					status: LowBalanceStatus.Low,
-					lastKnownAmount: 4000000n,
-					lastAlertedAt: null,
-				}),
-			}),
-		);
-		expect(mockLogWarn).not.toHaveBeenCalled();
-		expect(createdRule.status).toBe(LowBalanceStatus.Low);
-		expect(createdRule.lastAlertedAt).toBeNull();
-	});
-
-	it('resets a low rule and silently reseeds it healthy when the threshold is lowered', async () => {
-		mockUpdateLowBalanceRule.mockResolvedValue({
-			...createRuleRecord(LowBalanceStatus.Unknown),
-			thresholdAmount: 3000000n,
-		});
-		mockHotWalletFindFirst.mockResolvedValueOnce(createBalanceFetchWallet()).mockResolvedValueOnce({
-			...createWallet(LowBalanceStatus.Unknown),
-			LowBalanceRules: [
-				{
-					...createWallet(LowBalanceStatus.Unknown).LowBalanceRules[0],
-					thresholdAmount: 3000000n,
-				},
-			],
-		});
-		mockFetchAddressBalanceMap.mockResolvedValue(balanceMap(4000000n));
-		mockFindUniqueLowBalanceRule.mockResolvedValue({
-			...createRuleRecord(LowBalanceStatus.Healthy),
-			thresholdAmount: 3000000n,
-			lastKnownAmount: 4000000n,
-			lastCheckedAt: new Date('2026-03-10T12:01:00.000Z'),
-			lastAlertedAt: null,
-		});
-
-		const updatedRule = await service.updateRule({
-			ruleId: 'rule-1',
-			thresholdAmount: 3000000n,
-		});
-
-		expect(mockUpdateLowBalanceRule).toHaveBeenCalledWith(
-			expect.objectContaining({
-				data: expect.objectContaining({
-					thresholdAmount: 3000000n,
-					status: LowBalanceStatus.Unknown,
-					lastKnownAmount: null,
-					lastCheckedAt: null,
-					lastAlertedAt: null,
-				}),
-			}),
-		);
-		expect(mockUpdateMany).toHaveBeenCalledWith(
-			expect.objectContaining({
-				where: expect.objectContaining({
-					id: 'rule-1',
-					status: LowBalanceStatus.Unknown,
-				}),
-				data: expect.objectContaining({
-					status: LowBalanceStatus.Healthy,
-					lastKnownAmount: 4000000n,
-					lastAlertedAt: null,
-				}),
-			}),
-		);
-		expect(mockLogWarn).not.toHaveBeenCalled();
-		expect(updatedRule.status).toBe(LowBalanceStatus.Healthy);
-		expect(updatedRule.lastAlertedAt).toBeNull();
-	});
-
-	it('disables a rule without re-querying the wallet balance', async () => {
-		mockUpdateLowBalanceRule.mockResolvedValue(createRuleRecord(LowBalanceStatus.Unknown, false));
-		mockFindUniqueLowBalanceRule.mockResolvedValue(createRuleRecord(LowBalanceStatus.Unknown, false));
-
-		const updatedRule = await service.updateRule({
-			ruleId: 'rule-1',
-			enabled: false,
-		});
-
-		expect(mockFetchAddressBalanceMap).not.toHaveBeenCalled();
-		expect(mockHotWalletFindFirst).not.toHaveBeenCalled();
-		expect(updatedRule.enabled).toBe(false);
-		expect(updatedRule.status).toBe(LowBalanceStatus.Unknown);
-	});
-
-	it('retires an unclaimed top-up atomically when a rule changes', async () => {
-		mockUpdateLowBalanceRule.mockResolvedValue(createRuleRecord(LowBalanceStatus.Unknown, false));
-		mockFindUniqueLowBalanceRule.mockResolvedValue(createRuleRecord(LowBalanceStatus.Unknown, false));
-
-		await service.updateRule({
-			ruleId: 'rule-1',
-			enabled: false,
-		});
-
-		expect(mockDistributionUpdateMany).toHaveBeenCalledWith({
-			where: {
-				targetWalletId: 'wallet-1',
-				assetUnit: 'lovelace',
-				status: 'Pending',
-				fundWalletId: null,
-				transactionId: null,
-			},
-			data: {
-				status: 'Failed',
-				error: 'Distribution cancelled because its low-balance rule changed',
-			},
-		});
-	});
-
-	it('deletes a rule and retires its unclaimed top-up in one transaction', async () => {
-		mockDeleteLowBalanceRule.mockResolvedValue({
-			hotWalletId: 'wallet-1',
-			assetUnit: 'lovelace',
-		});
-
-		await service.deleteRule('rule-1');
-
-		expect(mockDeleteLowBalanceRule).toHaveBeenCalledWith({
-			where: { id: 'rule-1' },
-			select: {
-				hotWalletId: true,
-				assetUnit: true,
-			},
-		});
-		expect(mockDistributionUpdateMany).toHaveBeenCalledWith(
-			expect.objectContaining({
-				where: expect.objectContaining({
-					targetWalletId: 'wallet-1',
-					assetUnit: 'lovelace',
-					fundWalletId: null,
-					transactionId: null,
-				}),
-				data: expect.objectContaining({
-					status: 'Failed',
-					error: 'Distribution cancelled because its low-balance rule was deleted',
-				}),
-			}),
-		);
-	});
-
-	it('re-enables a rule and silently seeds low state without alerting', async () => {
-		mockUpdateLowBalanceRule.mockResolvedValue(createRuleRecord(LowBalanceStatus.Unknown));
-		mockHotWalletFindFirst
-			.mockResolvedValueOnce(createBalanceFetchWallet())
-			.mockResolvedValueOnce(createWallet(LowBalanceStatus.Unknown));
-		mockFetchAddressBalanceMap.mockResolvedValue(balanceMap(2000000n));
-		mockFindUniqueLowBalanceRule.mockResolvedValue({
-			...createRuleRecord(LowBalanceStatus.Low),
-			lastKnownAmount: 2000000n,
-			lastCheckedAt: new Date('2026-03-10T12:02:00.000Z'),
-			lastAlertedAt: null,
-		});
-
-		const updatedRule = await service.updateRule({
-			ruleId: 'rule-1',
-			enabled: true,
-		});
-
-		expect(mockFetchAddressBalanceMap).toHaveBeenCalledWith({
-			network: Network.Preprod,
-			rpcProviderApiKey: 'provider-key',
-			address: 'addr_test1...',
-		});
-		expect(mockLogWarn).not.toHaveBeenCalled();
-		expect(updatedRule.status).toBe(LowBalanceStatus.Low);
-		expect(updatedRule.lastAlertedAt).toBeNull();
 	});
 
 	it('logs scheduler failures and emits a summary after processing wallets', async () => {
