+++ src/services/wallets/low-balance/service.ts
@@ -1,24 +1,21 @@
 import { trace } from '@opentelemetry/api';
-import { Address, Transaction, Value } from '@emurgo/cardano-serialization-lib-nodejs';
-import type { UTxO } from '@meshsdk/core';
-import {
-	FundDistributionStatus,
-	HotWalletType,
-	Network,
-	PaymentSourceType,
-	type Prisma,
-} from '@/generated/prisma/client';
+import { HotWalletType, Network, PaymentSourceType } from '@/generated/prisma/client';
 import { LowBalanceStatus } from '@/generated/prisma/enums';
 import { prisma } from '@masumi/payment-core/db';
 import { retryOnSerializationConflict } from '@masumi/payment-core/db-retry';
-import { CONFIG, type LowBalanceDefaultRule } from '@masumi/payment-core/config';
 import { logger } from '@masumi/payment-core/logger';
 import { logWarn } from '@/utils/logs';
 import { recordWalletLowBalanceAlert } from '@masumi/payment-core/metrics';
 import { webhookEventsService } from '@/services/webhooks';
 import { fetchAddressBalanceMap } from '@/services/shared/address-balance';
-
-type BalanceMap = Map<string, bigint>;
+import { projectBalanceMapFromUnsignedTx, type BalanceMap, type ProjectableWalletUtxo } from './balance-map';
+import {
+	createRuleForWallet as createLowBalanceRuleForWallet,
+	deleteRule as deleteLowBalanceRule,
+	seedDefaultRulesForWallets as seedDefaultLowBalanceRulesForWallets,
+	updateRule as updateLowBalanceRule,
+	type WalletLowBalanceRuleMutationRecord,
+} from './rule-mutations';
 
 type WalletLowBalanceRuleRecord = {
 	id: string;
@@ -62,31 +59,6 @@ type EvaluateWalletContextOptions = {
 	emitAlerts?: boolean;
 };
 
-type WalletLowBalanceRuleMutationRecord = WalletLowBalanceRuleRecord & {
-	hotWalletId: string;
-};
-
-type MeshLikeUtxo = {
-	output: Pick<UTxO['output'], 'amount'>;
-};
-
-type LucidLikeUtxo = {
-	assets: Record<string, bigint>;
-};
-
-type ProjectableMeshLikeUtxo = {
-	input: Pick<UTxO['input'], 'txHash' | 'outputIndex'>;
-	output: Pick<UTxO['output'], 'amount'>;
-};
-
-type ProjectableLucidLikeUtxo = {
-	txHash: string;
-	outputIndex: number;
-	assets: Record<string, bigint>;
-};
-
-type ProjectableWalletUtxo = ProjectableMeshLikeUtxo | ProjectableLucidLikeUtxo;
-
 type ProjectedSubmissionEvaluation = {
 	hotWalletId: string;
 	walletAddress: string;
@@ -99,186 +71,12 @@ type ProjectedSubmissionEvaluation = {
 const LOW_BALANCE_WARNING_EVENT = 'wallet.low_balance';
 const SCHEDULED_MONITORING_CONCURRENCY = 5;
 
-async function retireUnclaimedTopupForRule(
-	tx: Prisma.TransactionClient,
-	rule: { hotWalletId: string; assetUnit: string },
-	error: string,
-): Promise<void> {
-	await tx.fundDistributionRequest.updateMany({
-		where: {
-			targetWalletId: rule.hotWalletId,
-			assetUnit: rule.assetUnit,
-			status: FundDistributionStatus.Pending,
-			fundWalletId: null,
-			transactionId: null,
-		},
-		data: {
-			status: FundDistributionStatus.Failed,
-			error,
-		},
-	});
-}
-
 function describeCheckSource(checkSource: WalletBalanceCheckSource): string {
 	return checkSource === 'interval_check' ? 'interval check' : 'submission';
 }
 
 function statusForBalance(currentAmount: bigint, thresholdAmount: bigint): LowBalanceStatus {
 	return currentAmount < thresholdAmount ? LowBalanceStatus.Low : LowBalanceStatus.Healthy;
-}
-
-function addQuantity(balanceMap: BalanceMap, assetUnit: string, quantity: bigint) {
-	balanceMap.set(assetUnit, (balanceMap.get(assetUnit) ?? 0n) + quantity);
-}
-
-function subtractQuantity(balanceMap: BalanceMap, assetUnit: string, quantity: bigint) {
-	balanceMap.set(assetUnit, (balanceMap.get(assetUnit) ?? 0n) - quantity);
-}
-
-function toBalanceMapFromMeshUtxos(utxos: MeshLikeUtxo[]): BalanceMap {
-	const balanceMap = new Map<string, bigint>();
-
-	for (const utxo of utxos) {
-		for (const amount of utxo.output.amount) {
-			const assetUnit = amount.unit === '' ? 'lovelace' : amount.unit;
-			addQuantity(balanceMap, assetUnit, BigInt(amount.quantity));
-		}
-	}
-
-	return balanceMap;
-}
-
-function toBalanceMapFromLucidUtxos(utxos: LucidLikeUtxo[]): BalanceMap {
-	const balanceMap = new Map<string, bigint>();
-
-	for (const utxo of utxos) {
-		for (const [assetUnit, quantity] of Object.entries(utxo.assets)) {
-			addQuantity(balanceMap, assetUnit === '' ? 'lovelace' : assetUnit, quantity);
-		}
-	}
-
-	return balanceMap;
-}
-
-function isLucidProjectableUtxo(utxo: ProjectableWalletUtxo): utxo is ProjectableLucidLikeUtxo {
-	return 'assets' in utxo;
-}
-
-function createUtxoReferenceKey(txHash: string, outputIndex: number): string {
-	return `${txHash}#${outputIndex}`;
-}
-
-function toBalanceMapFromProjectableUtxo(utxo: ProjectableWalletUtxo): BalanceMap {
-	if (isLucidProjectableUtxo(utxo)) {
-		return toBalanceMapFromLucidUtxos([utxo]);
-	}
-
-	return toBalanceMapFromMeshUtxos([utxo]);
-}
-
-function toBalanceMapFromProjectableUtxos(utxos: ProjectableWalletUtxo[]): BalanceMap {
-	const balanceMap = new Map<string, bigint>();
-
-	for (const utxo of utxos) {
-		const utxoBalanceMap = toBalanceMapFromProjectableUtxo(utxo);
-
-		for (const [assetUnit, quantity] of utxoBalanceMap) {
-			addQuantity(balanceMap, assetUnit, quantity);
-		}
-	}
-
-	return balanceMap;
-}
-
-function toBalanceMapFromCardanoValue(value: Value): BalanceMap {
-	const balanceMap = new Map<string, bigint>();
-
-	addQuantity(balanceMap, 'lovelace', BigInt(value.coin().to_str()));
-
-	const multiAsset = value.multiasset();
-	if (multiAsset == null) {
-		return balanceMap;
-	}
-
-	const policyIds = multiAsset.keys();
-	for (let policyIndex = 0; policyIndex < policyIds.len(); policyIndex++) {
-		const policyId = policyIds.get(policyIndex);
-		const assets = multiAsset.get(policyId);
-		if (assets == null) {
-			continue;
-		}
-
-		const assetNames = assets.keys();
-		for (let assetIndex = 0; assetIndex < assetNames.len(); assetIndex++) {
-			const assetName = assetNames.get(assetIndex);
-			const quantity = assets.get(assetName);
-			if (quantity == null) {
-				continue;
-			}
-
-			addQuantity(balanceMap, `${policyId.to_hex()}${assetName.to_hex()}`, BigInt(quantity.to_str()));
-		}
-	}
-
-	return balanceMap;
-}
-
-function projectBalanceMapFromUnsignedTx(
-	walletAddress: string,
-	walletUtxos: ProjectableWalletUtxo[],
-	unsignedTx: string,
-	currentBalanceMap?: BalanceMap,
-): BalanceMap {
-	const projectedBalanceMap =
-		currentBalanceMap == null ? toBalanceMapFromProjectableUtxos(walletUtxos) : new Map(currentBalanceMap);
-	const knownWalletInputs = new Map<string, BalanceMap>();
-
-	for (const utxo of walletUtxos) {
-		const referenceKey = isLucidProjectableUtxo(utxo)
-			? createUtxoReferenceKey(utxo.txHash, utxo.outputIndex)
-			: createUtxoReferenceKey(utxo.input.txHash, utxo.input.outputIndex);
-		knownWalletInputs.set(referenceKey, toBalanceMapFromProjectableUtxo(utxo));
-	}
-
-	const walletAddressHex = Address.from_bech32(walletAddress).to_hex();
-	const transaction = Transaction.from_bytes(Buffer.from(unsignedTx, 'hex'));
-	const transactionBody = transaction.body();
-	const inputs = transactionBody.inputs();
-
-	for (let inputIndex = 0; inputIndex < inputs.len(); inputIndex++) {
-		const input = inputs.get(inputIndex);
-		const inputBalanceMap = knownWalletInputs.get(
-			createUtxoReferenceKey(input.transaction_id().to_hex(), input.index()),
-		);
-
-		if (inputBalanceMap == null) {
-			continue;
-		}
-
-		for (const [assetUnit, quantity] of inputBalanceMap) {
-			subtractQuantity(projectedBalanceMap, assetUnit, quantity);
-		}
-	}
-
-	const outputs = transactionBody.outputs();
-	for (let outputIndex = 0; outputIndex < outputs.len(); outputIndex++) {
-		const output = outputs.get(outputIndex);
-		if (output.address().to_hex() !== walletAddressHex) {
-			continue;
-		}
-
-		for (const [assetUnit, quantity] of toBalanceMapFromCardanoValue(output.amount())) {
-			addQuantity(projectedBalanceMap, assetUnit, quantity);
-		}
-	}
-
-	return projectedBalanceMap;
-}
-
-function getNetworkDefaultLowBalanceRules(network: Network): LowBalanceDefaultRule[] {
-	return network === Network.Mainnet
-		? CONFIG.LOW_BALANCE_DEFAULT_RULES_MAINNET
-		: CONFIG.LOW_BALANCE_DEFAULT_RULES_PREPROD;
 }
 
 function serializeLowBalanceRecord(rule: WalletLowBalanceRuleRecord) {
@@ -353,43 +151,7 @@ export class WalletLowBalanceMonitorService {
 	}
 
 	async seedDefaultRulesForWallets(walletIds: string[]): Promise<void> {
-		if (walletIds.length === 0) {
-			return;
-		}
-
-		const wallets = await prisma.hotWallet.findMany({
-			where: {
-				id: { in: walletIds },
-				deletedAt: null,
-			},
-			select: {
-				id: true,
-				PaymentSource: {
-					select: {
-						network: true,
-					},
-				},
-			},
-		});
-
-		const rulesToCreate = wallets.flatMap((wallet) =>
-			getNetworkDefaultLowBalanceRules(wallet.PaymentSource.network).map((rule) => ({
-				hotWalletId: wallet.id,
-				assetUnit: rule.assetUnit,
-				thresholdAmount: BigInt(rule.thresholdAmount),
-				enabled: true,
-				status: LowBalanceStatus.Unknown,
-			})),
-		);
-
-		if (rulesToCreate.length === 0) {
-			return;
-		}
-
-		await prisma.hotWalletLowBalanceRule.createMany({
-			data: rulesToCreate,
-			skipDuplicates: true,
-		});
+		return seedDefaultLowBalanceRulesForWallets(walletIds);
 	}
 
 	async createRuleForWallet(params: {
@@ -400,37 +162,7 @@ export class WalletLowBalanceMonitorService {
 		topupEnabled?: boolean;
 		topupAmount?: bigint | null;
 	}): Promise<WalletLowBalanceRuleMutationRecord> {
-		const createdRule = await prisma.hotWalletLowBalanceRule.create({
-			data: {
-				hotWalletId: params.hotWalletId,
-				assetUnit: params.assetUnit,
-				thresholdAmount: params.thresholdAmount,
-				enabled: params.enabled,
-				topupEnabled: params.topupEnabled ?? false,
-				topupAmount: params.topupAmount ?? null,
-				status: LowBalanceStatus.Unknown,
-				lastKnownAmount: null,
-				lastCheckedAt: null,
-				lastAlertedAt: null,
-			},
-			select: {
-				id: true,
-				hotWalletId: true,
-				assetUnit: true,
-				thresholdAmount: true,
-				enabled: true,
-				topupEnabled: true,
-				topupAmount: true,
-				status: true,
-				lastKnownAmount: true,
-				lastCheckedAt: true,
-				lastAlertedAt: true,
-			},
-		});
-
-		await this.refreshRuleStateAfterMutation(createdRule.hotWalletId, createdRule.enabled);
-
-		return (await this.getRuleMutationRecordById(createdRule.id)) ?? createdRule;
+		return createLowBalanceRuleForWallet(this, params);
 	}
 
 	async updateRule(params: {
@@ -440,78 +172,11 @@ export class WalletLowBalanceMonitorService {
 		topupEnabled?: boolean;
 		topupAmount?: bigint | null;
 	}): Promise<WalletLowBalanceRuleMutationRecord> {
-		const updatedRule = await retryOnSerializationConflict(
-			() =>
-				prisma.$transaction(
-					async (tx) => {
-						const updated = await tx.hotWalletLowBalanceRule.update({
-							where: {
-								id: params.ruleId,
-							},
-							data: {
-								thresholdAmount: params.thresholdAmount,
-								enabled: params.enabled,
-								topupEnabled: params.topupEnabled,
-								topupAmount: params.topupAmount,
-								status: LowBalanceStatus.Unknown,
-								lastKnownAmount: null,
-								lastCheckedAt: null,
-								lastAlertedAt: null,
-							},
-							select: {
-								id: true,
-								hotWalletId: true,
-								assetUnit: true,
-								thresholdAmount: true,
-								enabled: true,
-								topupEnabled: true,
-								topupAmount: true,
-								status: true,
-								lastKnownAmount: true,
-								lastCheckedAt: true,
-								lastAlertedAt: true,
-							},
-						});
-
-						await retireUnclaimedTopupForRule(
-							tx,
-							updated,
-							'Distribution cancelled because its low-balance rule changed',
-						);
-						return updated;
-					},
-					{ isolationLevel: 'Serializable' },
-				),
-			{ label: 'low-balance-rule-update' },
-		);
-
-		await this.refreshRuleStateAfterMutation(updatedRule.hotWalletId, updatedRule.enabled);
-
-		return (await this.getRuleMutationRecordById(updatedRule.id)) ?? updatedRule;
+		return updateLowBalanceRule(this, params);
 	}
 
 	async deleteRule(ruleId: string): Promise<void> {
-		await retryOnSerializationConflict(
-			() =>
-				prisma.$transaction(
-					async (tx) => {
-						const deletedRule = await tx.hotWalletLowBalanceRule.delete({
-							where: { id: ruleId },
-							select: {
-								hotWalletId: true,
-								assetUnit: true,
-							},
-						});
-						await retireUnclaimedTopupForRule(
-							tx,
-							deletedRule,
-							'Distribution cancelled because its low-balance rule was deleted',
-						);
-					},
-					{ isolationLevel: 'Serializable' },
-				),
-			{ label: 'low-balance-rule-delete' },
-		);
+		return deleteLowBalanceRule(ruleId);
 	}
 
 	async evaluateHotWalletById(
@@ -874,46 +539,7 @@ export class WalletLowBalanceMonitorService {
 		return serializeLowBalanceSummary(rules);
 	}
 
-	private async getRuleMutationRecordById(ruleId: string): Promise<WalletLowBalanceRuleMutationRecord | null> {
-		return prisma.hotWalletLowBalanceRule.findUnique({
-			where: {
-				id: ruleId,
-			},
-			select: {
-				id: true,
-				hotWalletId: true,
-				assetUnit: true,
-				thresholdAmount: true,
-				enabled: true,
-				topupEnabled: true,
-				topupAmount: true,
-				status: true,
-				lastKnownAmount: true,
-				lastCheckedAt: true,
-				lastAlertedAt: true,
-			},
-		});
-	}
-
-	private async refreshRuleStateAfterMutation(hotWalletId: string, enabled: boolean): Promise<void> {
-		if (!enabled) {
-			return;
-		}
-
-		const balanceMap = await this.fetchCurrentBalanceMapForWallet(hotWalletId);
-		if (balanceMap == null) {
-			return;
-		}
-
-		const wallet = await this.getWalletLowBalanceContext(hotWalletId);
-		if (wallet == null || wallet.LowBalanceRules.length === 0) {
-			return;
-		}
-
-		await this.evaluateWalletContext(wallet, balanceMap, 'interval_check', { emitAlerts: false });
-	}
-
-	private async fetchCurrentBalanceMapForWallet(hotWalletId: string): Promise<BalanceMap | null> {
+	async fetchCurrentBalanceMapForWallet(hotWalletId: string): Promise<BalanceMap | null> {
 		const wallet = await prisma.hotWallet.findFirst({
 			where: {
 				id: hotWalletId,
@@ -973,14 +599,7 @@ export class WalletLowBalanceMonitorService {
 }
 
 export const walletLowBalanceMonitorService = new WalletLowBalanceMonitorService();
-export {
-	getNetworkDefaultLowBalanceRules,
-	projectBalanceMapFromUnsignedTx,
-	serializeLowBalanceRecord,
-	serializeLowBalanceSummary,
-	toBalanceMapFromLucidUtxos,
-	toBalanceMapFromMeshUtxos,
-};
+export { projectBalanceMapFromUnsignedTx, toBalanceMapFromLucidUtxos, toBalanceMapFromMeshUtxos } from './balance-map';
 export type {
 	BalanceMap,
 	LucidLikeUtxo,
@@ -988,6 +607,11 @@ export type {
 	ProjectableLucidLikeUtxo,
 	ProjectableMeshLikeUtxo,
 	ProjectableWalletUtxo,
+} from './balance-map';
+export { getNetworkDefaultLowBalanceRules } from './rule-mutations';
+export { serializeLowBalanceRecord, serializeLowBalanceSummary };
+export type {
+	EvaluateWalletContextOptions,
 	WalletBalanceCheckSource,
 	WalletLowBalanceContext,
 	WalletLowBalanceRuleRecord,
